added strand to server in prep for multi thread operation

+++ experiments/reader_test/mock_read_socket.cpp
@@ -25,7 +25,7 @@ MockReadSocket::~MockReadSocket()
     std::cout << __FUNCTION__ << std::endl;
     free((void*)_rdBuf);
 }
-int MockReadSocket::nativeSocketFD(){ return 9876; };
+long MockReadSocket::nativeSocketFD(){ return 9876; };
 void MockReadSocket::startRead()
 {
